requirements.txt
APScheduler==3.10.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
telebot==0.0.5
tzlocal==5.0.1
urllib3==2.0.4
